Moved lebdecode to binary utils file, added error message to failed custom token calls/
